requirements.txt
fastapi==0.109.0
pydantic==2.5.3
pydantic_core==2.14.6
pydyf==0.10.0
python-dotenv==1.1.1
resend==0.8.0
uvicorn==0.27.0
weasyprint==60.2
